Move selection to another page

+++ lib/data/tools/select_box.dart
@@ -36,8 +36,8 @@ class SelectBox extends Select {
     for (final stroke in selectResult.strokes) {
       colorDistribution.update(
         stroke.color,
-        (value) => value + stroke.length,
-        ifAbsent: () => stroke.length,
+        (value) => value + stroke.length as int,
+        ifAbsent: () => stroke.length as int,
       );
     }
     assert(colorDistribution.isNotEmpty);
